+++ server.js
@@ -3,7 +3,6 @@ const cors = require("cors");
 const app = express();
 const bodyParser = require("body-parser");
 const connectDB = require("./database.js");
-const FHB = require("./model/userData.js");
 require("dotenv").config();
 
 // RestfulAPIs
@@ -14,6 +13,10 @@ const saveFavs = require("./api/customize/saveFavs.js");
 const updateFavdetails = require("./api/customize/updateFavdet.js");
 const updateFavdesc = require("./api/customize/updateFavdesc.js");
 const deleteFav = require("./api/customize/deleteFav.js");
+const getOldbrew = require("./api/brewInfo/getOldbrews.js");
+const getRecentbrew = require("./api/brewInfo/getRecentbrew.js");
+const saveHistory = require("./api/brewInfo/saveHistory.js");
+const getFavourites = require("./api/brewInfo/getFavourites.js");
 
 connectDB();
 
@@ -49,76 +52,22 @@ app.get("/", (req, res) => {
 app.use(isAuthRouter);
 
 //GET history of oldBrews from DB
-app.get("/getOldbrews", async (req, res) => {
-  try {
-    const user = await FHB.findOne({ email: req.query.email });
-
-    res.send(user.oldBrews);
-  } catch (error) {
-    console.log(error);
-  }
-});
+app.use(getOldbrew);
 
 //GET Recent Brew history.
-app.get("/getRecentbrew", async (req, res) => {
-  try {
-    const user = await FHB.findOne({ email: req.query.email });
-
-    const sortedBrews = user.oldBrews.sort((a, b) => b.order - a.order);
-    const recentbrew = sortedBrews[0];
-
-    res.send(recentbrew);
-    console.log("Sent Recent Brew!");
-  } catch (error) {
-    console.log(error);
-  }
-});
+app.use(getRecentbrew);
 
 //GET favourites from DB
-app.get("/getFavourites", async (req, res) => {
-  try {
-    const user = await FHB.findOne({ email: req.query.email });
+app.use(getFavourites);
 
-    res.send(user.favourites);
-    console.log("Sent Favourites");
-  } catch (error) {
-    console.log(error);
-  }
-});
+//POST data oldBrews in databas
+app.use(saveHistory);
 
 // POST user data (register user)
 app.use(userRegister);
 
 // POST user data for login
 app.use(userLogIn);
-
-//POST data oldBrews in databas
-app.post("/saveHistory", async function (req, res) {
-  const userEmail = req.body.email;
-  const oldBrews = req.body.oldBrews[0];
-
-  try {
-    const user = await FHB.findOne({ email: userEmail });
-
-    if (user) {
-      if (user.oldBrews.length >= 50) {
-        // If the array has 10 or more elements, remove the oldest one
-        user.oldBrews.shift();
-      }
-
-      user.oldBrews.push(oldBrews);
-      await user.save();
-    } else {
-      res.status(404).send("Hisotry not saved!");
-    }
-
-    res.send(user.oldBrews);
-    console.log("Recent history Saved!");
-  } catch (error) {
-    console.log(error);
-    res.send("Server Error");
-  }
-});
 
 //POST favourite brews in database
 app.use(saveFavs);
